+++ docs/news.rst
@@ -4,10 +4,14 @@
 Relesase Notes
 ==============
 
-0.1.1 beta (xx-xx-2014)
-=======================
+0.2 alpha (xx-xx-2014)
+======================
 
 - added regex operator ``%`` for strings parsing
+- added multiline statements wrapped into parenthesis or square brackets
+- added inline comments
+- some HTTPS redirects are handled properly now
+- XPath refactoring and bug fixes
 
 0.1 pre-alpha (07-04-2014)
 ==========================